refactor: moved old events to new events dir
