+++ src/events/discordEvents/clientReady.ts
